divide report implementation
