+++ src/assault/report.rs
@@ -1,19 +1,21 @@
-use std::fmt::{Display, Formatter};
-use std::process::ExitCode;
+use std::{
+    fmt::{Display, Formatter},
+    io::{BufWriter, Stdout},
+    process::ExitCode,
+};
 
-use crate::error::Wrap;
 use crate::{
-    error::MultiWrap,
-    interface::command::{Relentless, ReportFormat},
-    interface::config::{
-        destinations::Destinations, http_serde_priv, Coalesce, Coalesced, Setting, Testcase, WorkerConfig,
+    error::{MultiWrap, Wrap},
+    interface::{
+        command::{Relentless, ReportFormat},
+        config::{destinations::Destinations, http_serde_priv, Coalesce, Coalesced, Setting, Testcase, WorkerConfig},
     },
 };
 
 /// TODO document
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct Report<T, Q, P> {
-    report: Vec<WorkerReport<T, Q, P>>,
+    pub report: Vec<WorkerReport<T, Q, P>>,
 }
 impl<T, Q: Clone + Coalesce, P: Clone + Coalesce> Report<T, Q, P> {
     pub fn new(report: Vec<WorkerReport<T, Q, P>>) -> Self {
@@ -32,8 +34,8 @@ impl<T, Q: Clone + Coalesce, P: Clone + Coalesce> Reportable for Report<T, Q, P>
 /// TODO document
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct WorkerReport<T, Q, P> {
-    config: Coalesced<WorkerConfig<Q, P>, Destinations<http_serde_priv::Uri>>,
-    report: Vec<CaseReport<T, Q, P>>,
+    pub config: Coalesced<WorkerConfig<Q, P>, Destinations<http_serde_priv::Uri>>,
+    pub report: Vec<CaseReport<T, Q, P>>,
 }
 impl<T, Q, P> WorkerReport<T, Q, P> {
     pub fn new(
@@ -52,9 +54,9 @@ impl<T, Q: Clone + Coalesce, P: Clone + Coalesce> Reportable for WorkerReport<T,
 /// TODO document
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct CaseReport<T, Q, P> {
-    testcase: Coalesced<Testcase<Q, P>, Setting<Q, P>>,
-    passed: usize,
-    messages: MultiWrap<T>,
+    pub testcase: Coalesced<Testcase<Q, P>, Setting<Q, P>>,
+    pub passed: usize,
+    pub messages: MultiWrap<T>,
 }
 impl<T, Q, P> CaseReport<T, Q, P> {
     pub fn new(testcase: Coalesced<Testcase<Q, P>, Setting<Q, P>>, passed: usize, messages: MultiWrap<T>) -> Self {
@@ -102,9 +104,9 @@ pub struct ReportWriter<W> {
     pub buf: W,
     pub at_start_line: bool,
 }
-impl ReportWriter<std::io::BufWriter<std::io::Stdout>> {
+impl ReportWriter<BufWriter<Stdout>> {
     pub fn with_stdout(indent: usize) -> Self {
-        let buf = std::io::BufWriter::new(std::io::stdout());
+        let buf = BufWriter::new(std::io::stdout());
         Self::new(indent, buf)
     }
 }
@@ -158,304 +160,5 @@ impl<W: std::io::Write> std::fmt::Write for ReportWriter<W> {
 impl<W: Display> Display for ReportWriter<W> {
     fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
         write!(f, "{}", self.buf)
-    }
-}
-
-#[cfg(feature = "console-report")]
-pub mod console_report {
-    use std::fmt::{Display, Write as _};
-
-    use crate::{
-        error::Wrap,
-        interface::command::Relentless,
-        interface::config::{Coalesce, Repeat, Testcase, WorkerConfig},
-    };
-
-    use super::{CaseReport, Report, ReportWriter, Reportable, WorkerReport};
-
-    pub trait ConsoleReport: Reportable {
-        type Error;
-        fn console_report<W: std::io::Write>(
-            &self,
-            cmd: &Relentless,
-            w: &mut ReportWriter<W>,
-        ) -> Result<(), Self::Error>;
-    }
-
-    impl<T: Display, Q: Clone + Coalesce, P: Clone + Coalesce> ConsoleReport for Report<T, Q, P> {
-        type Error = crate::Error;
-        fn console_report<W: std::io::Write>(
-            &self,
-            cmd: &Relentless,
-            w: &mut ReportWriter<W>,
-        ) -> Result<(), Self::Error> {
-            for report in &self.report {
-                if !report.skip_report(cmd) {
-                    report.console_report(cmd, w)?;
-                    writeln!(w).map_err(Wrap::wrapping)?;
-                }
-            }
-            Ok(())
-        }
-    }
-
-    pub enum WorkerConsoleReport {}
-    impl WorkerConsoleReport {
-        pub const NAME_DEFAULT: &'_ str = "testcases";
-        pub const NAME_EMOJI: console::Emoji<'_, '_> = console::Emoji("🚀", "");
-        pub const DESTINATION_EMOJI: console::Emoji<'_, '_> = console::Emoji("🌐", ":");
-        pub const OVERWRITE_DESTINATION_EMOJI: console::Emoji<'_, '_> = console::Emoji("👉", "->");
-    }
-
-    impl<T: Display, Q: Clone + Coalesce, P: Clone + Coalesce> ConsoleReport for WorkerReport<T, Q, P> {
-        type Error = Wrap; // TODO crate::Error ?
-        fn console_report<W: std::io::Write>(
-            &self,
-            cmd: &Relentless,
-            w: &mut ReportWriter<W>,
-        ) -> Result<(), Self::Error> {
-            let WorkerConfig { name, destinations, .. } = self.config.coalesce();
-
-            writeln!(
-                w,
-                "{} {} {}",
-                WorkerConsoleReport::NAME_EMOJI,
-                name.as_ref().unwrap_or(&WorkerConsoleReport::NAME_DEFAULT.to_string()),
-                WorkerConsoleReport::NAME_EMOJI
-            )?;
-
-            w.scope(|w| {
-                for (name, destination) in destinations {
-                    write!(w, "{}{} ", name, WorkerConsoleReport::DESTINATION_EMOJI)?;
-                    match self.config.base().destinations.get(&name) {
-                        Some(base) if base != &destination => {
-                            writeln!(
-                                w,
-                                "{} {} {}",
-                                **base,
-                                WorkerConsoleReport::OVERWRITE_DESTINATION_EMOJI,
-                                *destination
-                            )?;
-                        }
-                        _ => {
-                            writeln!(w, "{}", *destination)?;
-                        }
-                    }
-                }
-                Ok::<_, Wrap>(())
-            })?;
-
-            w.scope(|w| {
-                for report in &self.report {
-                    if !report.skip_report(cmd) {
-                        report.console_report(cmd, w)?;
-                    }
-                }
-                Ok::<_, Wrap>(())
-            })?;
-            Ok(())
-        }
-    }
-
-    pub enum CaseConsoleReport {}
-    impl CaseConsoleReport {
-        pub const PASS_EMOJI: console::Emoji<'_, '_> = console::Emoji("✅", "PASS");
-        pub const FAIL_EMOJI: console::Emoji<'_, '_> = console::Emoji("❌", "FAIL");
-        pub const REPEAT_EMOJI: console::Emoji<'_, '_> = console::Emoji("🔁", "");
-        pub const DESCRIPTION_EMOJI: console::Emoji<'_, '_> = console::Emoji("📝", "");
-        pub const ALLOW_EMOJI: console::Emoji<'_, '_> = console::Emoji("👀", "");
-        pub const MESSAGE_EMOJI: console::Emoji<'_, '_> = console::Emoji("💬", "");
-    }
-
-    impl<T: Display, Q: Clone + Coalesce, P: Clone + Coalesce> ConsoleReport for CaseReport<T, Q, P> {
-        type Error = Wrap; // TODO crate::Error ?
-        fn console_report<W: std::io::Write>(
-            &self,
-            cmd: &Relentless,
-            w: &mut ReportWriter<W>,
-        ) -> Result<(), Self::Error> {
-            let Testcase { description, target, setting, .. } = self.testcase.coalesce();
-
-            let side = if self.pass() { CaseConsoleReport::PASS_EMOJI } else { CaseConsoleReport::FAIL_EMOJI };
-            let target = console::style(&target);
-            write!(w, "{} {} ", side, if self.pass() { target.green() } else { target.red() })?;
-            if let Repeat(Some(ref repeat)) = setting.repeat {
-                write!(w, "{}{}/{} ", CaseConsoleReport::REPEAT_EMOJI, self.passed, repeat)?;
-            }
-            if let Some(ref description) = description {
-                writeln!(w, "{} {}", CaseConsoleReport::DESCRIPTION_EMOJI, description)?;
-            } else {
-                writeln!(w)?;
-            }
-            if !self.pass() && self.allow(cmd.strict) {
-                w.scope(|w| {
-                    writeln!(
-                        w,
-                        "{} {}",
-                        CaseConsoleReport::ALLOW_EMOJI,
-                        console::style("this testcase is allowed").green()
-                    )
-                })?;
-            }
-            if !self.messages.is_empty() {
-                w.scope(|w| {
-                    writeln!(
-                        w,
-                        "{} {}",
-                        CaseConsoleReport::MESSAGE_EMOJI,
-                        console::style("message was found").yellow()
-                    )?;
-                    w.scope(|w| {
-                        let message = &self.messages;
-                        writeln!(w, "{}", console::style(message).dim())
-                    })
-                })?;
-            }
-            Ok(())
-        }
-    }
-}
-
-pub mod github_markdown_report {
-    use std::fmt::{Display, Write as _};
-
-    use crate::{
-        error::Wrap,
-        interface::command::Relentless,
-        interface::config::{Coalesce, Repeat, Testcase, WorkerConfig},
-    };
-
-    use super::{CaseReport, Report, ReportWriter, Reportable, WorkerReport};
-
-    pub trait GithubMarkdownReport: Reportable {
-        type Error;
-        fn github_markdown_report<W: std::io::Write>(
-            &self,
-            cmd: &Relentless,
-            w: &mut ReportWriter<W>,
-        ) -> Result<(), Self::Error>;
-    }
-
-    impl<T: Display, Q: Clone + Coalesce, P: Clone + Coalesce> GithubMarkdownReport for Report<T, Q, P> {
-        type Error = crate::Error;
-        fn github_markdown_report<W: std::io::Write>(
-            &self,
-            cmd: &Relentless,
-            w: &mut ReportWriter<W>,
-        ) -> Result<(), Self::Error> {
-            for report in &self.report {
-                if !report.skip_report(cmd) {
-                    report.github_markdown_report(cmd, w)?;
-                    writeln!(w).map_err(Wrap::wrapping)?;
-                }
-            }
-            Ok(())
-        }
-    }
-
-    pub enum WorkerGithubMarkdownReport {}
-    impl WorkerGithubMarkdownReport {
-        pub const NAME_DEFAULT: &str = "testcases";
-        pub const NAME_EMOJI: &str = ":rocket:";
-        pub const DESTINATION_EMOJI: &str = ":globe_with_meridians:";
-        pub const OVERWRITE_DESTINATION_EMOJI: &str = ":point_right:";
-    }
-
-    impl<T: Display, Q: Clone + Coalesce, P: Clone + Coalesce> GithubMarkdownReport for WorkerReport<T, Q, P> {
-        type Error = Wrap; // TODO crate::Error ?
-        fn github_markdown_report<W: std::io::Write>(
-            &self,
-            cmd: &Relentless,
-            w: &mut ReportWriter<W>,
-        ) -> Result<(), Self::Error> {
-            let WorkerConfig { name, destinations, .. } = self.config.coalesce();
-
-            writeln!(
-                w,
-                "## {} {} {}",
-                WorkerGithubMarkdownReport::NAME_EMOJI,
-                name.as_ref().unwrap_or(&WorkerGithubMarkdownReport::NAME_DEFAULT.to_string()),
-                WorkerGithubMarkdownReport::NAME_EMOJI
-            )?;
-
-            for (name, destination) in destinations {
-                write!(w, "{} {} ", name, WorkerGithubMarkdownReport::DESTINATION_EMOJI)?;
-                match self.config.base().destinations.get(&name) {
-                    Some(base) if base != &destination => {
-                        writeln!(
-                            w,
-                            "{} {} {}",
-                            **base,
-                            WorkerGithubMarkdownReport::OVERWRITE_DESTINATION_EMOJI,
-                            *destination
-                        )?;
-                    }
-                    _ => {
-                        writeln!(w, "{}", *destination)?;
-                    }
-                }
-            }
-
-            for report in &self.report {
-                if !report.skip_report(cmd) {
-                    report.github_markdown_report(cmd, w)?;
-                }
-            }
-            Ok(())
-        }
-    }
-
-    pub enum CaseGithubMarkdownReport {}
-    impl CaseGithubMarkdownReport {
-        pub const PASS_EMOJI: &str = ":white_check_mark:";
-        pub const FAIL_EMOJI: &str = ":x:";
-        pub const REPEAT_EMOJI: &str = ":repeat:";
-        pub const DESCRIPTION_EMOJI: &str = ":memo:";
-        pub const ALLOW_EMOJI: &str = ":eyes:";
-        pub const MESSAGE_EMOJI: &str = ":speech_balloon:";
-    }
-
-    impl<T: Display, Q: Clone + Coalesce, P: Clone + Coalesce> GithubMarkdownReport for CaseReport<T, Q, P> {
-        type Error = Wrap; // TODO crate::Error ?
-        fn github_markdown_report<W: std::io::Write>(
-            &self,
-            cmd: &Relentless,
-            w: &mut ReportWriter<W>,
-        ) -> Result<(), Self::Error> {
-            let Testcase { description, target, setting, .. } = self.testcase.coalesce();
-
-            let side =
-                if self.pass() { CaseGithubMarkdownReport::PASS_EMOJI } else { CaseGithubMarkdownReport::FAIL_EMOJI };
-            write!(w, "- {} `{}` ", side, target)?;
-            if let Repeat(Some(ref repeat)) = setting.repeat {
-                write!(w, "{}{}/{} ", CaseGithubMarkdownReport::REPEAT_EMOJI, self.passed, repeat)?;
-            }
-            if let Some(ref description) = description {
-                writeln!(w, "{} {}", CaseGithubMarkdownReport::DESCRIPTION_EMOJI, description)?;
-            } else {
-                writeln!(w)?;
-            }
-            if !self.pass() && self.allow(cmd.strict) {
-                w.scope(|w| writeln!(w, "{} this testcase is allowed", CaseGithubMarkdownReport::ALLOW_EMOJI))?;
-            }
-            if !self.messages.is_empty() {
-                w.scope(|w| {
-                    writeln!(w, "<details>")?;
-                    w.scope(|w| {
-                        writeln!(
-                            w,
-                            "<summary> {} message was found </summary>",
-                            CaseGithubMarkdownReport::MESSAGE_EMOJI
-                        )?;
-                        writeln!(w)?;
-                        writeln!(w, "```")?;
-                        writeln!(w, "{}", &self.messages)?;
-                        writeln!(w, "```")
-                    })?;
-                    writeln!(w, "</details>")
-                })?;
-            }
-            Ok(())
-        }
     }
 }
