+++ src/interface/mod.rs
@@ -1,3 +1,4 @@
 pub mod command;
 pub mod config;
+pub mod report;
 pub mod template;
