+++ src/interface/report/mod.rs
@@ -1,0 +1,3 @@
+#[cfg(feature = "console-report")]
+pub mod console;
+pub mod github_markdown;
